requirements.txt
pandas==2.3.3
numpy==2.4.1
soccerdata==1.9.0
thefuzz==0.22.1
Unidecode==1.4.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic>=2.11.0,<3.0.0
sqlalchemy==2.0.36
psycopg2-binary>=2.9.10,<3.0.0
apscheduler==3.10.4
websockets>=15.0.1,<17.0.0
python-dotenv==1.0.0
